+++ services/product/src/main/java/com/jn/ecommerce/product/ProductPurchaseResponse.java
@@ -12,6 +12,6 @@ public record ProductPurchaseResponse(
 
         BigDecimal price,
 
-        double quantity
+        Integer quantity
 ) {
 }
